・HPControllerの記述をTargetControllerに統合。
