--- Assets/Scripts/HPController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5a779e6d257814df1859b8de76f34090
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
